chore: finish migrating editor tabs to trait impls, replaced scene bundle with requires
